--- a/.travis.yml
+++ b/.travis.yml
@@ -15,6 +15,19 @@ env:
   - TEST_SUITE=brakeman
 matrix:
   fast_finish: true
+  jobs:
+  - sudo: required
+    group: edge
+    virt: lxd
+    dist: bionic
+    arch: arm64-graviton2
+    rvm: 2.6.6
+    env:
+    - RUBY_GC_HEAP_GROWTH_MAX_SLOTS=300000
+    - RUBY_GC_HEAP_INIT_SLOTS=600000
+    - RUBY_GC_HEAP_GROWTH_FACTOR=1.25
+    - PARALLEL_TEST_PROCESSORS=2
+    - TEST_SUITE=vmdb PARALLEL=true
 addons:
   postgresql: '10'
 before_install: source ${TRAVIS_BUILD_DIR}/bin/ci/before_install.sh