--- a/bin/ci/before_install.sh
+++ b/bin/ci/before_install.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 set -v
 
+git config --global user.name "ManageIQ"
+git config --global user.email "contact@manageiq.org"
 source $TRAVIS_BUILD_DIR/bin/ci/setup_vmdb_configs.sh
 source $TRAVIS_BUILD_DIR/bin/ci/setup_ruby_env.sh