- Added ARM64 job in Travis-CI.
- Updated the git user details and the ip masking of 127.0.0.1 over 127.0.1.1.