@chessbyte 2020-11-30
Can you provide statistics for how much time each test suite takes before and after this PR?  Asking because this will run on EVERY commit of EVERY PR and we are already paying $$ to have extra Travis workers.