@odidev 2020-12-01
The arm64 is taking approximately equal time in comparison to the time taken by amd64.

For comparison I have taken the two builds under consideration:
Build1 : https://travis-ci.com/github/ManageIQ/manageiq/builds/205589190 ( Last master build before PR)
Build2: https://travis-ci.com/github/ManageIQ/manageiq/builds/203464013 (Our PR  build)

The following data is prepared while comparing the timings:


JOBS | AMD64 time before our PR | AMD64 total time after PR | ARM64 total time after PR | Difference of timing in amd64 before and after PR(Before PR - After PR)
-- | -- | -- | -- | --
rvm : 2.5.8, env : TEST_SUITE=vmdb PARALLEL=true | 1088 | 983 | 911 | 105
rvm : 2.5.8, env : TEST_SUITE=brakeman | 187 | 181 | 240 | 6
rvm : 2.6.6, env : TEST_SUITE=vmdb PARALLEL=true | 1029 | 1008 | 931 | 21
rvm : 2.6.6, env : TEST_SUITE=brakeman | 178 | 175 | 176 | 3
Total | 2482 | 2347 | 2258 | 135

As per above mentioned analysis we can say that  

Total Time taken by build After PR = Amd64 Total Time After PR + Arm64 Total Time After PR in seconds =  4605 sec
Total change in time in build will be (Total Time Taken after PR - Total Time taken before PR) = 2123 sec 
The total time taken by the builds for each commit will be increased by 2123 seconds .