@agrare 2021-02-02
@Fryguy this looks like it should be fine, cross-repo isn't looking at the matrix right now so as long as we have all the info in the rest of the file its going to work, just won't be able to run cross-repo arm tests.