@Fryguy 2021-02-02
Added a couple comments that I think we can reduce this down to just the core tests on Ruby 2.6 and still have confidence.  This would then only add 1 test suite worth of overhead, which may not be too bad.