@Fryguy 2021-02-04
This now only adds 1 extra run of 15 minutes (see https://travis-ci.com/github/ManageIQ/manageiq/jobs/480232961), which I'm good with.